Add code for propositional navigation
For a consistent user experience across services the proposition
navigation should be the consistent. This adds a hook to let the service
define the menu items and also adds the required css to make it look and
work the same as the existing ones on GOV.UK.

This uses querySelectorAll in the JavaScript because all browsers that
will understand the media query to display the mobile nav will
understand selectorAll.